requirements.txt
streamlit
faiss-cpu
numpy
openai
python-dotenv
